+++ macOS/Financer/Financer/Views/Finance/FinanceDetails.swift
